requirements.txt
fastapi
uvicorn
celery[redis]
twilio
redis
google-cloud-aiplatform
google-cloud-speech
requests
google-auth
python-dotenv
python-json-logger
pydantic-settings
